# performance optimizations & basic CICD for the submodule