--- a/.github/workflows/linux_builds_deployment.yml
+++ b/.github/workflows/linux_builds_deployment.yml
@@ -1,9 +1,8 @@
 name: 🐧 Linux Builds
-on: push
-# on: 
-#   push:
-#     paths:
-#       - godot-engine/**
+on: 
+  push:
+    paths:
+    - godot-engine/**
 
 defaults:
   run: