--- a/.github/workflows/windows_builds_deployment.yml
+++ b/.github/workflows/windows_builds_deployment.yml
@@ -1,9 +1,8 @@
 name: 🏁 Windows Builds
-on: push
-# on: 
-  # push:
-  #   paths:
-  #     - godot-engine/**
+on: 
+  push:
+    paths:
+    - godot-engine/**
 
 defaults:
   run: