--- a/mirror-godot-app/scripts/autoload/zone/client.gd
+++ b/mirror-godot-app/scripts/autoload/zone/client.gd
@@ -543,11 +543,9 @@ func _join_new_server_locally(space_id: String) -> bool:
 		if pid != null:
 			OS.kill(pid)
 		var firebase_auth = str(Firebase.Auth.auth.refreshtoken)
-		print(firebase_auth)
 		# For debugging this allows you to grab breakpoints from the server "--remote-debug", "tcp://127.0.0.1:6008"]
 		# If enabled it could cause join time to be much longer when booting server
-		var arguments = ["--server", "--space", space_id, "--mode", "edit", "--uuid", "localhost", "--server_login", firebase_auth, "--headless"]
-		print("SERVER ARGS: ", arguments)
+		var arguments = ["--server", "--space", space_id, "--mode", "edit", "--uuid", "localhost", "--server_login", firebase_auth, "--headless", "--remote-debug", "tcp://127.0.0.1:6008"]
 		pid = OS.create_process(OS.get_executable_path(), arguments, true)
 		start_join_localhost()
 		return true