--- a/mirror-godot-app/ui/game/connect_by_ip_ui.gd
+++ b/mirror-godot-app/ui/game/connect_by_ip_ui.gd
@@ -6,5 +6,5 @@ func _ready():
 
 func _on_connect_pressed():
 	var line_edit = $MarginContainer/HBoxContainer/LineEdit
-	print("Attempting connection: ", line_edit)
+	print("Attempting connection: ", line_edit.text)
 	Zone.client.connect_to_server_by_string(line_edit.text)