--- a/scripts/compile.sh
+++ b/scripts/compile.sh
@@ -1,3 +1,3 @@
 cd godot-engine
 # we enable a compile database so we can easily debug in vscode and or other tools
-scons target=editor optimize=speed_trace debug_symbols=yes compiledb=yes
+scons target=editor optimize=speed debug_symbols=no compiledb=yes production=true