What does this do?
- **Previous join time for the metro was was 1m 20s and is now down to 40 seconds.**
- **Join time for a fairly empty space is about 20-30 seconds.**
- **Server boot time reduced the new server boot time is 19s from 60 seconds with metro**
- **Prevents populating space objects multiple times**

Benchmarked/Tested on a CPU bound machine AMD Ryzen 5 4500

**TODO:**
- ~Client is disconnecting upon join this causes all instances to be reset, we should improve this so even if they reconnect they do not reset their objects until we're absolutely sure that is the correct action. (this should reduce the join time down a lot)~ (done by this PR, but can be improved) 
- Add an asset hash system to decrease load time considerably
- Shader compilation freezes cause network to disconnect, can we move this to another thread easily like the deep linking?